requirements.txt
hidapi
argparse